+++ src/components/App/App.css
@@ -73,3 +73,9 @@ Alert
   color: #e01616;
   font-size: 0.8rem;
 }
+
+@import url('https://fonts.googleapis.com/css2?family=Pattaya&display=swap');
+
+#header {
+    font-family: 'Pattaya', serif
+}
