+++ src/components/Search/appCurvyLines.png
