--- src/components/SearchView/SearchView.css
@@ -1,5 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Pattaya&display=swap');
-
-#header {
-    font-family: 'Pattaya', serif
-}
